requirements.txt
tensorflow-cpu==2.13.0
opencv-python==4.8.0.74
numpy==1.24.3
scikit-learn==1.3.0